+++ src/cli/CommandLineColors.ts
@@ -1,0 +1,6 @@
+export enum CommandLineColors {
+
+    RESET = "\x1b[0m",
+    RED = "\x1b[31m",
+    GREEN = "\x1b[32m"
+}