+++ src/target/lua/misc/LuaIdentifier.ts
@@ -1,13 +1,48 @@
 import { Target } from "../../Target";
 import * as ts from "typescript";
 import { BaseTarget } from "../../BaseTarget";
+import { LuaReservedKeywordError } from "../../../error/lua/LuaReservedKeywordError";
 
 export interface LuaIdentifier extends BaseTarget, Target { }
 export class LuaIdentifier implements Partial<Target> {
 
     public transpileIdentifier(node: ts.Identifier): string {
 
+        // check if the identifier is undefined
+        if (node.originalKeywordKind === ts.SyntaxKind.UndefinedKeyword) {
+            return this.transpileNode(ts.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword));
+        }
+
+        // get the identifier content
+        const identifier = node.escapedText.toString();
+
+        // check for reserved keywords
+        this.checkForReservedIdentifierToken(identifier, node);
+
         // get the identifier name
-        return node.escapedText.toString();
+        return identifier;
+    }
+
+    /**
+     * check if the identifier is a reserved lua keyword
+     * @param nodeContent the transpiled node content
+     */
+    private checkForReservedIdentifierToken(nodeContent: string, node: ts.Node): void {
+
+        // build a reserved keyword stack
+        // lua is case sensitive, so dont use a lower or upper case version
+        // of the identifier to compare
+        // @see http://www.lua.org/manual/5.1/manual.html#2.1
+        const reservedStack = [
+            "and", "break", "do", "else", "elseif",
+            "end", "false", "for", "function", "if",
+            "in", "local", "nil", "not", "or",
+            "repeat", "return", "then", "true", "until", "while"
+        ];
+
+        // make the test
+        if (reservedStack.indexOf(nodeContent) !== -1) {
+            throw new LuaReservedKeywordError(nodeContent, node);
+        }
     }
 }