+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 The qhun-transpiler project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Please consider every version below 1.0.0 as developement release and it may be unstable.
 
-## Unreleased
+## Todo for the next release
+
+- string.charAt()
+- string.split()
+- string.toLowerCase()
+- string.toUpperCase()
+
+## Implemented but unreleased
 
 - Added changelog file for future feature change tracking.
+- Added documentation for currently supported targets.
+- Added most of the typescript feature to `lua` and `wow` target